Updates to AtlasDatabaseUpdater.py ------------------------------------------------------------------------------- I changed this script to handle all of the collects we are using in the atlas database. It now has a list of collections that it knows about:
	Atlas_UseCases
        Atlas_activities
        Atlas_actors
        Atlas_cybersecurity_threats
        Atlas_disciplines
        Atlas_information_categories
        Atlas_information_types
        Atlas_locations
        Atlas_responding_organizations
        Atlas_technologies

In Export mode, this script now creates a JSON file for each of the
collections in this list in a directory specified by a command line
switch.

In Import mode, this script looks in the directory specified by the
command line switch for a JSON file for each of the above collections
(<collection name>.json)

Database Backup
-------------------------------------------------------------------------------
I added the directory atlas_database_backup that contains the JSON files
created with the new version of the AtlasDatabaseUpdater script.